Bump strava library temporarily
Strava access_token expires after six hours. There is code in the strava library
that automatically updates the access_token using the refresh_token. There is an
open PR that fixes the issue but hasn't been merged.

This is temporary until the PR has been merged into the main library.