Add csv export url for users with api_enabled
Adds a url like /exports/users/:slug/activities.csv which returns a csv of all
of the users activities in summary form. This includes distance, duration,
elevation, type, workout_type, and start_at_local.

The user must have api_enabled in their user_prefs in order to access this url.
This can be toggled on/off from /dashboard/settings/api

The goal of this is to use this csv information in observable, livebook, or
jupiter notebook.