Add bulk import script and upload url
Add the ability to upload a zip file of all the activity files in zip format
from a Strava bulk export.

This feature is only available in development right now. If it works well, maybe
we'll put it in production. The goal of this feature is to make it faster for
developers to play with their data.